refactor(odyssey): collapse formatOdysseyTime duplication
Move `formatOdysseyTime` from `@agent/odyssey/formatOdysseyTime.ts` into
`@tools/odyssey/odysseyMeta.ts` alongside `odysseyElapsedMs`. Both helpers
are now reachable from webview frontends through the existing
`@shared/schemas` re-export, so the OdysseyTab can drop its copy and use
the same canonical helper as the OdysseyTool. The tool view and the
settings tab can no longer drift on duration formatting.

(claude-bot review on da9e147)
